refactor : application 패키지 도메인별 그룹핑 통일
