+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/health/service/HealthService.kt
@@ -1,6 +1,6 @@
-package com.jaeyong.oop.application.service
+package com.jaeyong.oop.application.health.service
 
-import com.jaeyong.oop.application.usecase.HealthCheckUseCase
+import com.jaeyong.oop.application.health.usecase.HealthCheckUseCase
 import com.jaeyong.oop.domain.health.Health
 import com.jaeyong.oop.domain.health.port.HealthPort
 import org.springframework.stereotype.Service
@@ -8,12 +8,12 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class HealthService(
-    private val healthPort: HealthPort
+    private val healthRepository: HealthPort
 ) : HealthCheckUseCase {
     @Transactional
     override fun checkHealth(): String {
         val health = Health(status = "OK")
-        healthPort.save(health)
+        healthRepository.save(health)
         return "success"
     }
 }
