+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/health/usecase/HealthCheckUseCase.kt
@@ -1,4 +1,4 @@
-package com.jaeyong.oop.application.usecase
+package com.jaeyong.oop.application.health.usecase
 
 interface HealthCheckUseCase {
     fun checkHealth(): String
