+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/user/service/JoinService.kt
@@ -4,23 +4,24 @@ import com.jaeyong.oop.application.user.usecase.JoinUseCase
 import com.jaeyong.oop.common.exception.BaseException
 import com.jaeyong.oop.common.exception.ErrorCode
 import com.jaeyong.oop.domain.user.User
-import com.jaeyong.oop.domain.user.port.PasswordEncryptor
-import com.jaeyong.oop.domain.user.port.UserOutputPort
+import com.jaeyong.oop.application.user.common.JoinCommand
+import com.jaeyong.oop.domain.user.port.PasswordEncryptorPort
+import com.jaeyong.oop.domain.user.port.UserPort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class JoinService(
-    private val userRepository: UserOutputPort,
-    private val passwordEncryptor: PasswordEncryptor
+    private val userPort: UserPort,
+    private val passwordEncryptorPort: PasswordEncryptorPort
 ) : JoinUseCase {
 
     @Transactional
     override fun join(command: JoinCommand) {
-        if (userRepository.isUsernameTaken(command.username)) {
+        if (userPort.isUsernameTaken(command.username)) {
             throw BaseException(ErrorCode.DUPLICATE)
         }
-        val user = User.signUp(username = command.username, password = command.password, passwordEncryptor = passwordEncryptor)
-        userRepository.register(user)
+        val user = User.signUp(username = command.username, password = command.password, passwordEncryptor = passwordEncryptorPort)
+        userPort.register(user)
     }
 }
