+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/user/service/TokenValidationService.kt
@@ -1,0 +1,15 @@
+package com.jaeyong.oop.application.user.service
+
+import com.jaeyong.oop.application.user.usecase.TokenValidationUseCase
+import com.jaeyong.oop.domain.user.port.JwtHandlerPort
+import org.springframework.stereotype.Service
+
+@Service
+class TokenValidationService(
+    private val jwtHandlerPort: JwtHandlerPort
+) : TokenValidationUseCase {
+
+    override fun validateAndExtract(token: String): String? {
+        return jwtHandlerPort.validateAndExtract(token)
+    }
+}
