+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/user/usecase/JoinUseCase.kt
@@ -1,6 +1,6 @@
 package com.jaeyong.oop.application.user.usecase
 
-import com.jaeyong.oop.application.user.service.JoinCommand
+import com.jaeyong.oop.application.user.common.JoinCommand
 
 interface JoinUseCase {
     fun join(command: JoinCommand)
