+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/user/usecase/LoginUseCase.kt
@@ -1,4 +1,4 @@
-package com.jaeyong.oop.application.user.auth
+package com.jaeyong.oop.application.user.usecase
 
 interface LoginUseCase {
     fun login(username: String, password: String): String
