+++ oop-application/src/main/kotlin/com/jaeyong/oop/application/user/usecase/TokenValidationUseCase.kt
@@ -1,4 +1,4 @@
-package com.jaeyong.oop.application.user.auth
+package com.jaeyong.oop.application.user.usecase
 
 interface TokenValidationUseCase {
     fun validateAndExtract(token: String): String?
